feat: working on individual listing view
